# Optimize `mass_center` in `gymnasium/envs/mujoco/humanoid_v4.py`