# Description

<!-- CODEFLASH_OPTIMIZATION: {"function":"mass_center","file":"gymnasium/envs/mujoco/humanoid_v4.py","speedup_pct":"60%","speedup_x":"0.60x","original_runtime":"1.33 milliseconds","best_runtime":"832 microseconds","optimization_type":"memory","timestamp":"2025-08-06T22:23:29.823Z","version":"1.0"} -->
### 📄 60% (0.60x) speedup for ***`mass_center` in `gymnasium/envs/mujoco/humanoid_v4.py`***

⏱️ Runtime :   **`1.33 milliseconds`**  **→** **`832 microseconds`** (best of `164` runs)
### 📝 Explanation and details


The optimized code achieves a **60% speedup** by replacing inefficient array operations with `np.einsum` and eliminating unnecessary array expansions.

**Key optimizations:**

1. **Eliminated `np.expand_dims`**: The original code expanded `body_mass` from shape `(b,)` to `(b,1)` just for broadcasting, which creates an unnecessary intermediate array and memory allocation.

2. **Replaced broadcasting multiplication with `einsum`**: Instead of `mass * xpos` which creates a large intermediate array of shape `(b,j)`, `np.einsum('b,bj->j', model.body_mass, data.xipos)` performs the weighted sum directly without the intermediate array. This is more memory-efficient and computationally faster.

3. **Direct sum computation**: Changed `np.sum(mass)` to `model.body_mass.sum()`, avoiding the need to sum the expanded mass array.

**Why this is faster:**
- **Memory efficiency**: Eliminates intermediate arrays, reducing memory allocation overhead
- **Vectorized operations**: `einsum` is highly optimized for this exact pattern of weighted sums
- **Cache locality**: Better memory access patterns with fewer temporary arrays

**Performance gains by test case:**
- Small arrays (2-3 bodies): ~45-50% faster
- Medium arrays (1000 bodies): ~50-100% faster  
- Large-scale tests show the biggest gains (up to 113% faster) as the memory and computation benefits compound with array size

The optimization is particularly effective for larger arrays where memory allocation costs dominate, but provides consistent speedups across all test cases.



✅ **Correctness verification report:**

| Test                        | Status            |
| --------------------------- | ----------------- |
| ⚙️ Existing Unit Tests | 🔘 **None Found** |
| 🌀 Generated Regression Tests | ✅ **27 Passed** |
| ⏪ Replay Tests | 🔘 **None Found** |
| 🔎 Concolic Coverage Tests | 🔘 **None Found** |
|📊 Tests Coverage       | 100.0% |
<details>
<summary>🌀 Generated Regression Tests and Runtime</summary>

```python
import numpy as np
# imports
import pytest  # used for our unit tests
from gymnasium.envs.mujoco.humanoid_v4 import mass_center


# Helper classes to simulate 'model' and 'data'
class DummyModel:
    def __init__(self, body_mass):
        self.body_mass = body_mass

class DummyData:
    def __init__(self, xipos):
        self.xipos = xipos

# ---------------------------
# Basic Test Cases
# ---------------------------

def test_single_body_at_origin():
    # One body at the origin
    model = DummyModel(np.array([1.0]))
    data = DummyData(np.array([[0.0, 0.0, 0.0]]))
    codeflash_output = mass_center(model, data); center = codeflash_output # 43.2μs -> 29.3μs (47.2% faster)

def test_single_body_nonzero():
    # One body at a non-origin position
    model = DummyModel(np.array([2.5]))
    data = DummyData(np.array([[3.0, 4.0, 5.0]]))
    codeflash_output = mass_center(model, data); center = codeflash_output # 42.5μs -> 29.1μs (45.8% faster)

def test_two_bodies_equal_mass():
    # Two bodies, equal mass, at (0,0,0) and (2,2,2)
    model = DummyModel(np.array([1.0, 1.0]))
    data = DummyData(np.array([[0.0, 0.0, 0.0], [2.0, 2.0, 2.0]]))
    codeflash_output = mass_center(model, data); center = codeflash_output # 43.6μs -> 29.5μs (48.0% faster)

def test_two_bodies_unequal_mass():
    # Two bodies, mass 1 and 3, at (0,0,0) and (2,0,0)
    model = DummyModel(np.array([1.0, 3.0]))
    data = DummyData(np.array([[0.0, 0.0, 0.0], [2.0, 0.0, 0.0]]))
    # Center: (1*0 + 3*2) / 4 = 1.5
    codeflash_output = mass_center(model, data); center = codeflash_output # 43.5μs -> 29.1μs (49.3% faster)

def test_three_bodies_varied():
    # Three bodies, masses 1, 2, 3 at (0,0,0), (1,0,0), (0,2,0)
    model = DummyModel(np.array([1.0, 2.0, 3.0]))
    data = DummyData(np.array([
        [0.0, 0.0, 0.0],
        [1.0, 0.0, 0.0],
        [0.0, 2.0, 0.0]
    ]))
    # x: (1*0 + 2*1 + 3*0)/6 = 0.333..., y: (1*0 + 2*0 + 3*2)/6 = 1.0
    codeflash_output = mass_center(model, data); center = codeflash_output # 43.5μs -> 29.2μs (49.2% faster)

# ---------------------------
# Edge Test Cases
# ---------------------------



def test_zero_mass_and_nonzero_mass():
    # One zero-mass body, one nonzero-mass body
    model = DummyModel(np.array([0.0, 5.0]))
    data = DummyData(np.array([[100.0, 200.0, 300.0], [1.0, 2.0, 3.0]]))
    codeflash_output = mass_center(model, data); center = codeflash_output # 44.0μs -> 29.6μs (48.5% faster)

def test_all_bodies_same_position():
    # All bodies at same position, any masses
    model = DummyModel(np.array([1.0, 5.0, 3.0]))
    data = DummyData(np.array([[7.0, 8.0, 9.0], [7.0, 8.0, 9.0], [7.0, 8.0, 9.0]]))
    codeflash_output = mass_center(model, data); center = codeflash_output # 43.7μs -> 29.6μs (47.4% faster)

def test_large_and_small_masses():
    # One very large, one very small mass
    model = DummyModel(np.array([1e-9, 1e9]))
    data = DummyData(np.array([[100.0, 200.0, 300.0], [1.0, 2.0, 3.0]]))
    codeflash_output = mass_center(model, data); center = codeflash_output # 43.8μs -> 28.9μs (51.4% faster)

def test_high_dimensional_positions():
    # More than 3D positions, should use only first three
    model = DummyModel(np.array([1.0, 1.0]))
    data = DummyData(np.array([[1.0, 2.0, 3.0, 4.0], [5.0, 6.0, 7.0, 8.0]]))
    # Function only returns first two coordinates
    codeflash_output = mass_center(model, data); center = codeflash_output # 43.7μs -> 29.4μs (48.4% faster)
    expected = [ (1.0*1.0 + 1.0*5.0)/2, (1.0*2.0 + 1.0*6.0)/2 ]



def test_many_bodies_uniform():
    # 1000 bodies, all at (1,2,3), mass 1
    n = 1000
    model = DummyModel(np.ones(n))
    data = DummyData(np.tile([1.0, 2.0, 3.0], (n, 1)))
    codeflash_output = mass_center(model, data); center = codeflash_output # 68.4μs -> 36.2μs (88.8% faster)

def test_many_bodies_random_positions_and_masses():
    # 1000 bodies, random positions and masses
    np.random.seed(42)
    n = 1000
    masses = np.random.rand(n) * 10 + 1  # masses in [1, 11)
    positions = np.random.rand(n, 3) * 100  # positions in [0, 100)
    model = DummyModel(masses)
    data = DummyData(positions)
    # Compute expected center using numpy
    expected = np.sum(masses[:,None] * positions, axis=0) / np.sum(masses)
    codeflash_output = mass_center(model, data); center = codeflash_output # 53.9μs -> 26.8μs (101% faster)

def test_large_scale_performance():
    # 999 bodies, all at random positions, all mass 1
    np.random.seed(0)
    n = 999
    model = DummyModel(np.ones(n))
    data = DummyData(np.random.uniform(-1e3, 1e3, size=(n,3)))
    expected = np.mean(data.xipos, axis=0)
    codeflash_output = mass_center(model, data); center = codeflash_output # 59.2μs -> 27.8μs (113% faster)

def test_large_scale_extreme_values():
    # 1000 bodies, some with very large/small masses and positions
    np.random.seed(123)
    n = 1000
    masses = np.random.lognormal(mean=0, sigma=5, size=n)
    positions = np.random.normal(loc=0, scale=1e6, size=(n,3))
    model = DummyModel(masses)
    data = DummyData(positions)
    expected = np.sum(masses[:,None] * positions, axis=0) / np.sum(masses)
    codeflash_output = mass_center(model, data); center = codeflash_output # 52.5μs -> 26.1μs (101% faster)
# codeflash_output is used to check that the output of the original code is the same as that of the optimized code.
#------------------------------------------------
import numpy as np
# imports
import pytest  # used for our unit tests
from gymnasium.envs.mujoco.humanoid_v4 import mass_center


# Helper classes to simulate 'model' and 'data'
class DummyModel:
    def __init__(self, body_mass):
        self.body_mass = np.array(body_mass)

class DummyData:
    def __init__(self, xipos):
        self.xipos = np.array(xipos)

# ------------------------
# Basic Test Cases
# ------------------------

def test_single_body_at_origin():
    # One body, mass 1, at origin
    model = DummyModel([1.0])
    data = DummyData([[0.0, 0.0, 0.0]])
    codeflash_output = mass_center(model, data); result = codeflash_output # 43.8μs -> 30.2μs (45.3% faster)

def test_two_bodies_equal_mass_opposite_sides():
    # Two bodies, equal mass, positions symmetric about origin
    model = DummyModel([2.0, 2.0])
    data = DummyData([[1.0, 0.0, 0.0],
                      [-1.0, 0.0, 0.0]])
    codeflash_output = mass_center(model, data); result = codeflash_output # 44.2μs -> 29.7μs (48.7% faster)

def test_two_bodies_unequal_mass():
    # Two bodies, different masses
    model = DummyModel([1.0, 3.0])
    data = DummyData([[0.0, 0.0, 0.0],
                      [2.0, 0.0, 0.0]])
    codeflash_output = mass_center(model, data); result = codeflash_output # 43.1μs -> 29.2μs (47.8% faster)

def test_three_bodies_arbitrary_positions():
    # Three bodies, arbitrary positions and masses
    model = DummyModel([1.0, 2.0, 3.0])
    data = DummyData([[0.0, 1.0, 0.0],
                      [2.0, 0.0, 0.0],
                      [1.0, 1.0, 1.0]])
    # Center of mass calculation:
    # x: (1*0 + 2*2 + 3*1) / 6 = (0+4+3)/6 = 7/6
    # y: (1*1 + 2*0 + 3*1) / 6 = (1+0+3)/6 = 4/6 = 2/3
    codeflash_output = mass_center(model, data); result = codeflash_output # 43.6μs -> 29.4μs (48.2% faster)
    expected = [7/6, 2/3]

# ------------------------
# Edge Test Cases
# ------------------------


def test_single_body_nonzero_position():
    # One body, arbitrary nonzero position
    model = DummyModel([42.0])
    data = DummyData([[7.0, -3.0, 5.0]])
    codeflash_output = mass_center(model, data); result = codeflash_output # 45.0μs -> 31.1μs (44.8% faster)


def test_all_bodies_same_position():
    # All bodies at the same position, different masses
    model = DummyModel([1.0, 2.0, 3.0])
    data = DummyData([[5.0, 5.0, 5.0],
                      [5.0, 5.0, 5.0],
                      [5.0, 5.0, 5.0]])
    codeflash_output = mass_center(model, data); result = codeflash_output # 46.3μs -> 31.8μs (45.5% faster)

def test_large_and_small_masses():
    # One body dominates the mass
    model = DummyModel([1e-9, 1e9])
    data = DummyData([[1000.0, 1000.0, 1000.0],
                      [2.0, 3.0, 4.0]])
    # Center of mass should be very close to [2.0, 3.0]
    codeflash_output = mass_center(model, data); result = codeflash_output # 44.6μs -> 29.3μs (52.1% faster)

def test_high_dimensional_positions():
    # Positions have more than 3 dimensions (should work for any n>=3)
    model = DummyModel([1.0, 2.0])
    data = DummyData([[1.0, 2.0, 3.0, 4.0],
                      [4.0, 3.0, 2.0, 1.0]])
    # Only first two dimensions are returned
    codeflash_output = mass_center(model, data); result = codeflash_output # 44.0μs -> 29.7μs (48.2% faster)
    # (1*1 + 2*4)/(1+2)= (1+8)/3=3, (1*2+2*3)/(1+2)=(2+6)/3=8/3
    expected = [3.0, 8/3]

def test_non_float_masses_and_positions():
    # Integer masses and positions
    model = DummyModel([1, 2, 3])
    data = DummyData([[0, 0, 0],
                      [2, 0, 0],
                      [1, 1, 1]])
    codeflash_output = mass_center(model, data); result = codeflash_output # 49.2μs -> 33.8μs (45.9% faster)
    expected = [7/6, 2/3]

# ------------------------
# Large Scale Test Cases
# ------------------------

def test_many_bodies_random_positions():
    # Test with 1000 bodies at random positions and masses
    np.random.seed(42)
    n = 1000
    masses = np.random.rand(n)
    positions = np.random.rand(n, 3)
    model = DummyModel(masses)
    data = DummyData(positions)
    # Compute expected result
    total_mass = np.sum(masses)
    expected = np.sum(masses[:, None] * positions, axis=0) / total_mass
    expected = expected[:2]
    codeflash_output = mass_center(model, data); result = codeflash_output # 52.1μs -> 26.3μs (98.0% faster)

def test_many_bodies_same_position():
    # All bodies at the same position, large n
    n = 1000
    masses = np.arange(1, n+1)
    pos = [7.7, -8.8, 9.9]
    positions = np.tile(pos, (n, 1))
    model = DummyModel(masses)
    data = DummyData(positions)
    codeflash_output = mass_center(model, data); result = codeflash_output # 69.6μs -> 46.7μs (49.0% faster)

def test_many_bodies_one_dominant_mass():
    # 999 small masses at random, one huge mass at a fixed position
    np.random.seed(0)
    n = 1000
    masses = np.ones(n)
    positions = np.random.randn(n, 3)
    # Set last mass much larger and at a known position
    masses[-1] = 1e6
    positions[-1] = [10.0, 20.0, 30.0]
    model = DummyModel(masses)
    data = DummyData(positions)
    codeflash_output = mass_center(model, data); result = codeflash_output # 67.7μs -> 34.9μs (93.7% faster)

def test_large_scale_performance():
    # This test is not for correctness but to ensure it runs in reasonable time
    n = 999
    masses = np.random.rand(n)
    positions = np.random.rand(n, 3)
    model = DummyModel(masses)
    data = DummyData(positions)
    codeflash_output = mass_center(model, data); result = codeflash_output # 68.9μs -> 37.1μs (85.8% faster)
# codeflash_output is used to check that the output of the original code is the same as that of the optimized code.
```

</details>


To edit these changes `git checkout codeflash/optimize-mass_center-me0jbcsc` and push.

[![Codeflash](https://img.shields.io/badge/Optimized%20with-Codeflash-yellow?style=flat&color=%23ffc428&logo=data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNDgwIiBoZWlnaHQ9ImF1dG8iIHZpZXdCb3g9IjAgMCA0ODAgMjgwIiBmaWxsPSJub25lIiB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciPgo8cGF0aCBmaWxsLXJ1bGU9ImV2ZW5vZGQiIGNsaXAtcnVsZT0iZXZlbm9kZCIgZD0iTTI4Ni43IDAuMzc4NDE4SDIwMS43NTFMNTAuOTAxIDE0OC45MTFIMTM1Ljg1MUwwLjk2MDkzOCAyODEuOTk5SDk1LjQzNTJMMjgyLjMyNCA4OS45NjE2SDE5Ni4zNDVMMjg2LjcgMC4zNzg0MThaIiBmaWxsPSIjRkZDMDQzIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMzExLjYwNyAwLjM3ODkwNkwyNTguNTc4IDU0Ljk1MjZIMzc5LjU2N0w0MzIuMzM5IDAuMzc4OTA2SDMxMS42MDdaIiBmaWxsPSIjMEIwQTBBIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMzA5LjU0NyA4OS45NjAxTDI1Ni41MTggMTQ0LjI3NkgzNzcuNTA2TDQzMC4wMjEgODkuNzAyNkgzMDkuNTQ3Vjg5Ljk2MDFaIiBmaWxsPSIjMEIwQTBBIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMjQyLjg3MyAxNjQuNjZMMTg5Ljg0NCAyMTkuMjM0SDMxMC44MzNMMzYzLjM0NyAxNjQuNjZIMjQyLjg3M1oiIGZpbGw9IiMwQjBBMEEiLz4KPC9zdmc+Cg==)](https://codeflash.ai)

Fixes # (issue)

## Type of change

- [x] New feature (non-breaking change which adds functionality)

# Checklist:

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with `pre-commit run --all-files` (see `CONTRIBUTING.md` instructions to set it up)
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes