@Kallinteris-Andreas 2025-08-26
1) Why apply this change to only Humanoid-v4 and not Humanoid-v5?
2) I think that `mass_center()` now needs a docstring (or maybe I am the only one who cannot intuitively read `np.einsum`)
P.S. I run a test for 100k random steps, and `assert mass_center_old() = mass_center_new()` passed.